feat: final

+++ app/Http/Controllers/PatientsController.php
@@ -2,16 +2,20 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\NewMedicineEvent;
 use App\Events\PatientAddedEvent;
 use App\Events\PatientEvent;
 use App\Models\Disease;
 use App\Models\History;
 use App\Models\LabResults;
+use App\Models\Medicine;
 use App\Models\Note;
 use App\Models\Patients;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Yajra\DataTables\DataTables;
+
+use function Sodium\compare;
 
 class PatientsController extends Controller
 {
@@ -39,7 +43,13 @@ class PatientsController extends Controller
 
     
     public function show(Patients $patients){
-        return view('admin.'.$this->folderName.'.detail', compact('patients'));
+        $medicine = Medicine::all();
+        $total = History::where('name', $patients->name)->count();
+        return view('admin.'.$this->folderName.'.detail', compact('patients', 'medicine', 'total'));
+    }
+
+    public function edit(Patients $patients){
+        return view('admin.'.$this->folderName.'.form', compact('patients'));
     }
 
     public function store(Request $request){
@@ -48,6 +58,20 @@ class PatientsController extends Controller
             Patients::create($request->all());
             event(new PatientAddedEvent());
             $msg = 'Patient Berhasil Ditambah';
+        } catch (\Exception $e){
+            $msg = $e->getMessage();
+        }
+
+        return redirect('/admin/patient')->with('message', $msg);
+
+    }
+
+    public function update(Request $request, Patients $patients){
+        
+        try {
+            Patients::find($patients->id)->update($request->all());
+            event(new PatientAddedEvent());
+            $msg = 'Patient Berhasil Diubah';
         } catch (\Exception $e){
             $msg = $e->getMessage();
         }
@@ -73,15 +97,19 @@ class PatientsController extends Controller
     public function createMedicine(Request $request,Patients $patients){
 
         try {
-
-            History::create($request->all());
+            History::create([...$request->all(), 'name' => $patients->name]);
             $msg = 'Obat Berhasil Ditambah';
-            event(new PatientEvent());
+            event(new NewMedicineEvent());
         } catch (\Exception $e){
             $msg = $e->getMessage();
         }
 
         return redirect('/admin/patient/show/'. $patients->id)->with('message', $msg);
+    }
+
+    public function showAllMedicine(Patients $patients){
+        $medicine = Medicine::all();
+        return view('admin.'.$this->folderName.'.allHistory', compact('patients', 'medicine'));
     }
 
     public function updateMedicine(Request $request, Patients $patients, History $history){
@@ -116,7 +144,7 @@ class PatientsController extends Controller
     public function createDiseaseHistory(Request $request,Patients $patients){
 
         try {
-            Disease::create($request->all());
+            Disease::create([...$request->all(), 'name' => $patients->name]);
             $msg = 'Riawayat Penyakit Berhasil Ditambah';
             event(new PatientEvent());
         } catch (\Exception $e){
@@ -161,7 +189,8 @@ class PatientsController extends Controller
         try {
             LabResults::create([
                 'patients_id' => $patients->id,
-                'image' => $img
+                'image' => $img,
+                'note' => $request->note
             ]);
             event(new PatientEvent());
             $msg = 'Foto Berhasil Ditambah';
@@ -175,11 +204,18 @@ class PatientsController extends Controller
 
     public function updateLabPhoto(Request $request, Patients $patients, LabResults $labResults){
 
-        Storage::delete($labResults->image);
-        $img = $request->file('image')->store('lab-result');
+        $data = $request->validate([
+            'note' => 'required',
+            'image' => 'nullable'
+        ]);
+
+        if($request->image){
+            Storage::delete($labResults->image);
+            $data['image'] = $request->file('image')->store('lab-result');
+        }
 
         try {
-            LabResults::find($labResults->id)->update(['image' => $img]);
+            LabResults::find($labResults->id)->update($data);
             event(new PatientEvent());
             $msg = 'Foto Berhasil Diubah';
         } catch (\Exception $e){
